Remove feature `convenience`
This feature was originally designed as an opt-out from panicking methods, which have a Result/Option-based
overload, such as from_variant (try_from_variant), instance_id (instance_id_or_none), cast (try_cast), etc.

However, it does not really add much in practice besides maintenance overhead. To still warn people to not
accidentally use one of these methods, a warning sign is prepended to the brief doc. This should show up
first in IDEs.

Note that the warning sign is not added to every panicking method, but particularly those which have a "try"
overload. To know which methods panic, check out if it contains a `# Panics` section.